requirements.txt
fastapi==0.115.12
uvicorn==0.34.2
SQLAlchemy==2.0.40
asyncpg==0.30.0
python-dotenv==1.1.0
email_validator==2.2.0
Jinja2==3.1.6
python-multipart==0.0.20
python-socketio==5.13.0
python-engineio==4.12.1
aiohttp==3.11.18
